Move tests to dedicated files
