+++ glean-core/tests/simple.rs
@@ -180,87 +180,3 @@ fn transformation_works() {
         format!("metrics snapshot 2: {}", metrics_snapshot)
     );
 }
-
-#[test]
-fn uuid() {
-    let (_t, tmpname) = tempdir();
-    let mut glean = Glean::new(&tmpname, GLOBAL_APPLICATION_ID).unwrap();
-
-    let uuid: UuidMetric = UuidMetric::new(CommonMetricData {
-        name: "uuid".into(),
-        category: "local".into(),
-        send_in_pings: vec!["core".into()],
-        ..Default::default()
-    });
-
-    uuid.generate(&glean);
-    let snapshot = glean.snapshot("core", false);
-    assert!(
-        snapshot.contains(r#""local.uuid": ""#),
-        format!("Snapshot 1: {}", snapshot)
-    );
-
-    uuid.generate(&glean);
-    let snapshot = glean.snapshot("core", false);
-    assert!(
-        snapshot.contains(r#""local.uuid": ""#),
-        format!("Snapshot 2: {}", snapshot)
-    );
-}
-
-#[test]
-fn list() {
-    let (_t, tmpname) = tempdir();
-    let mut glean = Glean::new(&tmpname, GLOBAL_APPLICATION_ID).unwrap();
-
-    let list: StringListMetric = StringListMetric::new(CommonMetricData {
-        name: "list".into(),
-        category: "local".into(),
-        send_in_pings: vec!["core".into()],
-        ..Default::default()
-    });
-
-    list.add(&glean, "first");
-    let snapshot = glean.snapshot("core", false);
-    assert!(snapshot.contains(r#""local.list": ["#));
-    assert!(snapshot.contains(r#""first""#));
-
-    list.add(&glean, "second");
-    let snapshot = glean.snapshot("core", false);
-    assert!(snapshot.contains(r#""local.list": ["#));
-    assert!(snapshot.contains(r#""first""#));
-    assert!(snapshot.contains(r#""second""#));
-
-    list.set(&glean, vec!["third".into()]);
-    let snapshot = glean.snapshot("core", false);
-    assert!(snapshot.contains(r#""local.list": ["#));
-    assert!(!snapshot.contains(r#""first""#));
-    assert!(!snapshot.contains(r#""second""#));
-    assert!(snapshot.contains(r#""third""#));
-}
-
-#[test]
-fn write_ping_to_disk() {
-    let (temp, tmpname) = tempdir();
-    let glean = Glean::new(&tmpname, GLOBAL_APPLICATION_ID).unwrap();
-
-    // We need to store a metric as an empty ping is not stored.
-    let counter = CounterMetric::new(CommonMetricData {
-        name: "counter".into(),
-        category: "local".into(),
-        send_in_pings: vec!["metrics".into()],
-        ..Default::default()
-    });
-    counter.add(&glean, 1);
-
-    assert!(glean.send_ping("metrics", false).unwrap());
-
-    let path = temp.path().join("pings");
-
-    let mut count = 0;
-    for entry in std::fs::read_dir(path).unwrap() {
-        assert!(entry.unwrap().path().is_file());
-        count += 1;
-    }
-    assert_eq!(1, count);
-}
